Time: 10 ms (79.55%), Space: 46.4 MB (40.91%) - LeetHub

+++ 1390-four-divisors/1390-four-divisors.java
@@ -18,10 +18,10 @@ class Solution {
         return c==4?sum:0;
     }
     public int sumFourDivisors(int[] nums) {
-        long ans = 0;
+        int ans = 0;
         for(int i : nums){
             ans += four(i);
         }
-        return (int)ans;
+        return ans;
     }
 }
